refactor: group app-slot code in a single folder
