+++ src/components/generic/app-slot/AppSlotContent.vue
@@ -1,6 +1,6 @@
 <script>
 import { onActivated, onBeforeUnmount, onDeactivated, onMounted } from "vue";
-import { useAppSlot } from "@/composables/app-slot";
+import { useAppSlot } from "./app-slot";
 
 export default {
   props: {
